+++ modules/bridge/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- New `slider` dialog component #TINY-8304
+- New `spacer` dialog component #TINY-8304
+- New `buttonType` property on dialog button components, supporting `toolbar` style in addition to `primary` and `secondary` #TINY-8304
+
+### Deprecated
+- The dialog button component `primary` property has been deprecated in favour of the new `buttonType` property #TINY-8304
+
 ### Changed
 - Upgraded to Katamari 9.0, which includes breaking changes to the `Optional` API used in this module.